Engrave preview: Properly cancel rendering threads
Before, Thread.stop() was used, which is deprecated and doesn't really work. This causes the problem that when quickly resizing / changing the piece, multiple ImageProcessingThreads keep running and each uses RAM.

Now, we use Thread.interrupt() and InterruptedException. With this, threads quit voluntarily and not immediately. Therefore we have to wait for previous threads to exit, so that only one rendering buffer is in RAM at a given time.