(semi-)Automatic update for laser settings
Every 14 days, ask if new laser settings should be downloaded. This can be disabled from preferences and is automatically disabled if the preferences are from an older VisiCut version, were loaded from a local file or are a git repository.

To avoid the need for provisioning (#375), auto-update is enabled by default for hosts of FAU FabLab (and, in the future, other labs which are auto-detected by hostname FQDN -- see MainView.getRecommendedLab() ), even if the preferences are from an older VisiCut version.

(also includes some improvements of the German translation)