Fix location of embedded JRE in macOS bundle
This was inadvertently broken in 4bd60b0. The JRE is intended to be
installed under Contents/Plugins/JRE, but the refactoring inadvertently
made it be installed under Contents/Plugins only. In turn, this meant
the launched script could no longer find the JRE and gave an error
instead of starting the application.

Fixes #684.