Preserve the user entered text  in the JobName TextField (#687)
Preserve the value of the usr entered JobName TextField.

Currently visicut deletes the value, when a) the laser is changed, b) material is changed,
or c) the mapping is changed.
This makes it tedious, if I want to e.g. laser the same file different mappings, which I frequently do.
I'd prefer to just edit the prefix, instead of entering it again in full each time.